requirements.txt
fastapi
uvicorn
PyPDF2
psycopg2-binary
